+++ cmd/ingest_metrics.go
@@ -22,6 +22,7 @@ import (
 	"github.com/spf13/cobra"
 	"go.opentelemetry.io/otel/attribute"
 
+	"github.com/cardinalhq/lakerunner/config"
 	"github.com/cardinalhq/lakerunner/internal/awsclient"
 	"github.com/cardinalhq/lakerunner/internal/debugging"
 	"github.com/cardinalhq/lakerunner/internal/healthcheck"
@@ -91,5 +92,9 @@ func init() {
 
 func metricIngestBatch(ctx context.Context, ll *slog.Logger, tmpdir string, sp storageprofile.StorageProfileProvider, mdb lrdb.StoreFull,
 	awsmanager *awsclient.Manager, items []lrdb.Inqueue, ingest_dateint int32, rpfEstimate int64, loop *IngestLoopContext) error {
-	return ingestion.ProcessBatch(ctx, ll, tmpdir, sp, mdb, awsmanager, items, ingest_dateint, rpfEstimate, loop.exemplarProcessor)
+	cfg, err := config.Load()
+	if err != nil {
+		return err
+	}
+	return ingestion.ProcessBatch(ctx, ll, tmpdir, sp, mdb, awsmanager, items, ingest_dateint, rpfEstimate, loop.exemplarProcessor, cfg.Metrics.Ingestion)
 }